Revert "[6.0] Introduce Joomla 6.0 compatibility plugin (#45336)"
This reverts commit cb6cb1d494d283d87032f4f93919aa2d5a4da376.
