+++ administrator/components/com_admin/script.php
@@ -287,8 +287,6 @@ class JoomlaInstallerScript
              * ['type' => 'plugin', 'element' => 'demotasks', 'folder' => 'task', 'client_id' => 0, 'pre_function' => null],
              * ['type' => 'plugin', 'element' => 'compat', 'folder' => 'system', 'client_id' => 0, 'pre_function' => 'migrateCompatPlugin'],
              */
-            // Uninstall the old joomla 5 compatibility plugin
-            ['type' => 'plugin', 'element' => 'compat', 'folder' => 'system', 'client_id' => 0, 'pre_function' => null],
         ];
 
         $db = Factory::getDbo();
