+++ administrator/language/en-GB/plg_behaviour_compat.ini
@@ -1,0 +1,15 @@
+; Joomla! Project
+; (C) 2023 Open Source Matters, Inc. <https://www.joomla.org>
+; License GNU General Public License version 2 or later; see LICENSE.txt
+; Note : All ini files need to be saved as UTF-8
+
+PLG_BEHAVIOUR_COMPAT="Behaviour - Backward Compatibility"
+PLG_COMPAT_FIELD_CLASSES_ALIASES_LABEL="Classes Aliases"
+PLG_COMPAT_FIELD_CLASSES_ALIASES_DESCRIPTION="Add class aliases for classes which have been renamed or moved to a namespace."
+PLG_COMPAT_FIELD_ES5_ASSETS_DESCRIPTION="Activate this option if your extension requires *.es5 assets which has resulted in an exception. The assets provided are empty but prevent the exception."
+PLG_COMPAT_FIELD_ES5_ASSETS_LABEL="ES5 Assets"
+PLG_COMPAT_FIELD_LEGACY_CLASSES_DESCRIPTION="Activate this option if your extension requires classes deprecated for the current version of Joomla. This includes those removed classes from the folder /plugins/behavior/compat/classes/."
+PLG_COMPAT_FIELD_LEGACY_CLASSES_LABEL="Include Deprecated Classes"
+PLG_COMPAT_FIELD_REMOVED_ASSETS_DESCRIPTION="Activate this option if your extension requires removed assets which has resulted in an exception. The assets provided are empty but prevent the exception."
+PLG_COMPAT_FIELD_REMOVED_ASSETS_LABEL="Removed Assets"
+PLG_COMPAT_XML_DESCRIPTION="If you use extensions which are not using the current Joomla Coding standards then this plugin, when enabled, will provide backward compatibility to the prior major version."
