+++ administrator/language/en-GB/plg_behaviour_compat.sys.ini
@@ -1,0 +1,7 @@
+; Joomla! Project
+; (C) 2023 Open Source Matters, Inc. <https://www.joomla.org>
+; License GNU General Public License version 2 or later; see LICENSE.txt
+; Note : All ini files need to be saved as UTF-8
+
+PLG_BEHAVIOUR_COMPAT="Behaviour - Backward Compatibility"
+PLG_COMPAT_XML_DESCRIPTION="Provides backward compatibility to the prior major version."
