+++ administrator/language/en-GB/plg_behaviour_compat6.ini
@@ -4,10 +4,4 @@
 ; Note : All ini files need to be saved as UTF-8
 
 PLG_BEHAVIOUR_COMPAT6="Behaviour - Backward Compatibility 6"
-PLG_COMPAT6_FIELD_CLASSES_ALIASES_DESCRIPTION="Add class aliases for classes which have been renamed or moved to a namespace."
-PLG_COMPAT6_FIELD_CLASSES_ALIASES_LABEL="Classes Aliases"
-PLG_COMPAT6_FIELD_LEGACY_CLASSES_DESCRIPTION="Activate this option if your extension requires classes deprecated for the current version of Joomla. This includes those removed classes from the folder /plugins/behavior/compat6/classes/."
-PLG_COMPAT6_FIELD_LEGACY_CLASSES_LABEL="Include Deprecated Classes"
-PLG_COMPAT6_FIELD_REMOVED_ASSETS_DESCRIPTION="Activate this option if your extension requires removed assets which has resulted in an exception. The assets provided are empty but prevent the exception."
-PLG_COMPAT6_FIELD_REMOVED_ASSETS_LABEL="Removed Assets"
-PLG_COMPAT6_XML_DESCRIPTION="If you use extensions which are not using the current Joomla Coding standards then this plugin, when enabled, will provide backward compatibility to the previous major version, Joomla 5."
+PLG_COMPAT6_XML_DESCRIPTION="This plugin should be enabled before updating to Joomla 6."
