+++ administrator/language/en-GB/plg_behaviour_compat6.sys.ini
@@ -4,4 +4,4 @@
 ; Note : All ini files need to be saved as UTF-8
 
 PLG_BEHAVIOUR_COMPAT6="Behaviour - Backward Compatibility 6"
-PLG_COMPAT6_XML_DESCRIPTION="If you use extensions which are not using the current Joomla Coding standards then this plugin, when enabled, will provide backward compatibility to the previous major version, Joomla 5."
+PLG_COMPAT6_XML_DESCRIPTION="This plugin should be enabled before updating to Joomla 6."
