+++ plugins/behaviour/compat/classes/Application/CLI/ColorStyle.php
