+++ plugins/behaviour/compat/classes/Application/CLI/Output/Processor/ColorProcessor.php
