+++ plugins/behaviour/compat/classes/Application/CLI/Output/Processor/ProcessorInterface.php
