+++ plugins/behaviour/compat/classes/Application/CLI/Output/Stdout.php
