+++ plugins/behaviour/compat/classes/Application/CLI/Output/Xml.php
