+++ plugins/behaviour/compat/classes/Filesystem/Stream.php
