+++ plugins/behaviour/compat/classes/Filesystem/Streams/StreamString.php
