+++ plugins/behaviour/compat/classes/Filesystem/Support/StringController.php
