+++ plugins/behaviour/compat/compat.xml
@@ -1,0 +1,76 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<extension type="plugin" group="behaviour" method="upgrade">
+	<name>plg_behaviour_compat</name>
+	<author>Joomla! Project</author>
+	<creationDate>2023-09</creationDate>
+	<copyright>(C) 2023 Open Source Matters, Inc.</copyright>
+	<license>GNU General Public License version 2 or later; see LICENSE.txt</license>
+	<authorEmail>admin@joomla.org</authorEmail>
+	<authorUrl>www.joomla.org</authorUrl>
+	<version>5.0.0</version>
+	<description>PLG_COMPAT_XML_DESCRIPTION</description>
+	<namespace path="src">Joomla\Plugin\Behaviour\Compat</namespace>
+	<files>
+		<folder>classes</folder>
+		<folder plugin="compat">services</folder>
+		<folder>src</folder>
+	</files>
+	<languages>
+		<language tag="en-GB">language/en-GB/plg_behaviour_compat.ini</language>
+		<language tag="en-GB">language/en-GB/plg_behaviour_compat.sys.ini</language>
+	</languages>
+	<config>
+		<fields name="params">
+			<fieldset name="basic">
+				<field
+					name="classes_aliases"
+					type="radio"
+					layout="joomla.form.field.radio.switcher"
+					label="PLG_COMPAT_FIELD_CLASSES_ALIASES_LABEL"
+					description="PLG_COMPAT_FIELD_CLASSES_ALIASES_DESCRIPTION"
+					default="1"
+					filter="integer"
+					>
+					<option value="0">JNO</option>
+					<option value="1">JYES</option>
+				</field>
+				<field
+					name="legacy_classes"
+					type="radio"
+					layout="joomla.form.field.radio.switcher"
+					label="PLG_COMPAT_FIELD_LEGACY_CLASSES_LABEL"
+					description="PLG_COMPAT_FIELD_LEGACY_CLASSES_DESCRIPTION"
+					default="1"
+					filter="integer"
+					>
+					<option value="0">JNO</option>
+					<option value="1">JYES</option>
+				</field>
+				<field
+					name="es5_assets"
+					type="radio"
+					label="PLG_COMPAT_FIELD_ES5_ASSETS_LABEL"
+					description="PLG_COMPAT_FIELD_ES5_ASSETS_DESCRIPTION"
+					layout="joomla.form.field.radio.switcher"
+					default="1"
+					filter="integer"
+					>
+					<option value="0">JNO</option>
+					<option value="1">JYES</option>
+				</field>
+				<field
+					name="removed_asset"
+					type="radio"
+					label="PLG_COMPAT_FIELD_REMOVED_ASSETS_LABEL"
+					description="PLG_COMPAT_FIELD_REMOVED_ASSETS_DESCRIPTION"
+					layout="joomla.form.field.radio.switcher"
+					default="1"
+					filter="integer"
+					>
+					<option value="0">JNO</option>
+					<option value="1">JYES</option>
+				</field>
+			</fieldset>
+		</fields>
+	</config>
+</extension>
