+++ plugins/behaviour/compat6/compat6.xml
@@ -7,11 +7,10 @@
 	<license>GNU General Public License version 2 or later; see LICENSE.txt</license>
 	<authorEmail>admin@joomla.org</authorEmail>
 	<authorUrl>www.joomla.org</authorUrl>
-	<version>6.0.0</version>
+	<version>5.4.0</version>
 	<description>PLG_COMPAT6_XML_DESCRIPTION</description>
 	<namespace path="src">Joomla\Plugin\Behaviour\Compat6</namespace>
 	<files>
-		<folder>classes</folder>
 		<folder plugin="compat6">services</folder>
 		<folder>src</folder>
 	</files>
@@ -19,35 +18,4 @@
 		<language tag="en-GB">language/en-GB/plg_behaviour_compat6.ini</language>
 		<language tag="en-GB">language/en-GB/plg_behaviour_compat6.sys.ini</language>
 	</languages>
-	<config>
-		<fields name="params">
-			<fieldset name="basic">
-				<field
-					name="classes_aliases"
-					type="radio"
-					label="PLG_COMPAT6_FIELD_CLASSES_ALIASES_LABEL"
-					description="PLG_COMPAT6_FIELD_CLASSES_ALIASES_DESCRIPTION"
-					layout="joomla.form.field.radio.switcher"
-					default="1"
-					filter="integer"
-					>
-					<option value="0">JNO</option>
-					<option value="1">JYES</option>
-				</field>
-				<field
-					name="legacy_classes"
-					type="radio"
-					label="PLG_COMPAT6_FIELD_LEGACY_CLASSES_LABEL"
-					description="PLG_COMPAT6_FIELD_LEGACY_CLASSES_DESCRIPTION"
-					layout="joomla.form.field.radio.switcher"
-					default="1"
-					filter="integer"
-					>
-					<option value="0">JNO</option>
-					<option value="1">JYES</option>
-				</field>
-
-			</fieldset>
-		</fields>
-	</config>
 </extension>
